refactor: move kia api types into kia subdirectory
