+++ KiaMaps/Core/Api/ApiConfiguration.swift
@@ -105,159 +105,6 @@ protocol ApiConfiguration {
     var apiProviderKind: ApiProviderKind { get }
 }
 
-/// European region API configuration for supported vehicle brands
-/// Provides brand-specific endpoints, credentials, and service identifiers for EU market
-enum KiaApiConfigurationEurope: String, ApiConfiguration {
-    case kia
-    case hyundai
-    case genesis
-
-    var key: String {
-        switch self {
-        case .kia:
-            "kia"
-        case .hyundai:
-            "hyundai"
-        case .genesis:
-            "genesis"
-        }
-    }
-
-    var name: String {
-        switch self {
-        case .kia:
-            "Kia"
-        case .hyundai:
-            "Hyundai"
-        case .genesis:
-            "Genesis"
-        }
-    }
-
-    var port: Int {
-        switch self {
-        case .kia:
-            8080
-        case .hyundai, .genesis:
-            443
-        }
-    }
-
-    var serviceAgent: String {
-        "okhttp/3.12.0"
-    }
-
-    var userAgent: String {
-        let device = UIDevice.current
-        return "EU_BlueLink/2.1.18 (com.kia.connect.eu; build:10560; \(device.systemName) \(device.systemVersion)) Alamofire/5.8.0"
-    }
-
-    var acceptHeader: String {
-        "text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9"
-    }
-
-    var baseHost: String {
-        switch self {
-        case .kia:
-            "https://prd.eu-ccapi.kia.com"
-        case .hyundai:
-            "https://prd.eu-ccapi.hyundai.com"
-        case .genesis:
-            "https://prd-eu-ccapi.genesis.com"
-        }
-    }
-
-    var loginHost: String {
-        "https://idpconnect-eu.\(key).com"
-    }
-
-    var mqttHost: String {
-        "https://egw-svchub-ccs-\(brandCode.lowercased())-eu.eu-central.hmgmobility.com:31010"
-    }
-
-    var serviceId: String {
-        switch self {
-        case .kia:
-            "fdc85c00-0a2f-4c64-bcb4-2cfb1500730a"
-        case .hyundai:
-            "6d477c38-3ca4-4cf3-9557-2a1929a94654"
-        case .genesis:
-            "3020afa2-30ff-412a-aa51-d28fbe901e10"
-        }
-    }
-
-    var appId: String {
-        switch self {
-        case .kia:
-            "a2b8469b-30a3-4361-8e13-6fceea8fbe74"
-        case .hyundai:
-            "014d2225-8495-4735-812d-2616334fd15d"
-        case .genesis:
-            "f11f2b86-e0e7-4851-90df-5600b01d8b70"
-        }
-    }
-
-    var senderId: Int {
-        199_360_397_125
-    }
-
-    var authClientId: String {
-        switch self {
-        case .kia:
-            "572e0304-5f8d-4b4c-9dd5-41aa84eed160"
-        case .hyundai:
-            "64621b96-0f0d-11ec-82a8-0242ac130003"
-        case .genesis:
-            "3020afa2-30ff-412a-aa51-d28fbe901e10"
-        }
-    }
-
-    var cfb: String {
-        switch self {
-        case .kia:
-            "wLTVxwidmH8CfJYBWSnHD6E0huk0ozdiuygB4hLkM5XCgzAL1Dk5sE36d/bx5PFMbZs="
-        case .hyundai:
-            "RFtoRq/vDXJmRndoZaZQyfOot7OrIqGVFj96iY2WL3yyH5Z/pUvlUhqmCxD2t+D65SQ="
-        case .genesis:
-            "RFtoRq/vDXJmRndoZaZQyYo3/qFLtVReW8P7utRPcc0ZxOzOELm9mexvviBk/qqIp4A="
-        }
-    }
-
-    var brandCode: String {
-        switch self {
-        case .kia:
-            "K"
-        case .hyundai:
-            "H"
-        case .genesis:
-            "G"
-        }
-    }
-
-    var brandName: String {
-        switch self {
-        case .kia:
-            "Kia"
-        case .hyundai:
-            "Hyundai"
-        case .genesis:
-            "Genesis"
-        }
-    }
-
-    var pushType: String {
-        if self == .kia {
-            "APNS"
-        } else {
-            "GCM"
-        }
-    }
-
-    var apiProviderKind: ApiProviderKind {
-        .hmg
-    }
-}
-
 enum PorscheApiConfiguration: String, ApiConfiguration {
     case europe
     case usa
